2.0/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pyglet==2.0.15
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
ttkthemes==3.2.2
tzdata==2024.1
urllib3==2.2.1